Hente ut pid fra idPortenToken
