+++ frontend/arrangor-flate/package.json
@@ -23,11 +23,11 @@
     "@navikt/ds-react": "^7.0.0",
     "@navikt/ds-tailwind": "^7.0.0",
     "@navikt/nav-dekoratoren-moduler": "^3.1.0",
+    "@navikt/oasis": "^3.4.0",
     "@remix-run/express": "^2.12.1",
     "@remix-run/node": "^2.12.1",
     "@remix-run/react": "^2.12.1",
     "@remix-run/serve": "^2.12.1",
-    "@navikt/oasis": "^3.4.0",
     "compression": "^1.7.4",
     "cross-env": "^7.0.3",
     "csp-header": "^5.2.1",
@@ -51,8 +51,10 @@
     "@tailwindcss/typography": "^0.5.15",
     "@types/compression": "^1.7.5",
     "@types/express": "^4.17.21",
+    "@types/morgan": "^1.9.9",
     "@types/react": "^18.3.8",
     "@types/react-dom": "^18.3.0",
+    "@types/uuid": "^10.0.0",
     "@typescript-eslint/eslint-plugin": "^8.6.0",
     "@typescript-eslint/parser": "^8.6.0",
     "autoprefixer": "^10.4.20",
